refactor(richtext): Move RichTextDataProcessor
Moving `RichTextDataProcessor`, so that we may have a legacy
and a new API, chosen at higher level.
