+++ packages/ckeditor5-coremedia-richtext/src/index-doc.ts
@@ -18,8 +18,8 @@ export { default as CoreMediaRichText } from "./CoreMediaRichText";
 export * from "./CoreMediaRichTextConfig";
 export { default as CoreMediaRichTextConfig } from "./CoreMediaRichTextConfig";
 
-export * from "./RichTextDataProcessor";
-export { default as RichTextDataProcessor } from "./RichTextDataProcessor";
+export * from "./processors/RichTextDataProcessor";
+export { default as RichTextDataProcessor } from "./processors/RichTextDataProcessor";
 
 export * from "./RichTextSchema";
 export { default as RichTextSchema } from "./RichTextSchema";
