+++ packages/ckeditor5-coremedia-richtext/src/processors/RichTextDataProcessor.ts
@@ -6,15 +6,15 @@ import { MatcherPattern } from "@ckeditor/ckeditor5-engine/src/view/matcher";
 import Logger from "@coremedia/ckeditor5-logging/logging/Logger";
 import LoggerProvider from "@coremedia/ckeditor5-logging/logging/LoggerProvider";
 import DomConverter from "@ckeditor/ckeditor5-engine/src/view/domconverter";
-import RichTextXmlWriter from "./RichTextXmlWriter";
+import RichTextXmlWriter from "../RichTextXmlWriter";
 import HtmlFilter from "@coremedia/ckeditor5-dataprocessor-support/HtmlFilter";
-import RichTextSchema from "./RichTextSchema";
-import { COREMEDIA_RICHTEXT_PLUGIN_NAME } from "./Constants";
+import RichTextSchema from "../RichTextSchema";
+import { COREMEDIA_RICHTEXT_PLUGIN_NAME } from "../Constants";
 import Editor from "@ckeditor/ckeditor5-core/src/editor/editor";
-import { getConfig } from "./CoreMediaRichTextConfig";
+import { getConfig } from "../CoreMediaRichTextConfig";
 import { HtmlWriter } from "@ckeditor/ckeditor5-engine/src/dataprocessor/htmlwriter";
 import BasicHtmlWriter from "@ckeditor/ckeditor5-engine/src/dataprocessor/basichtmlwriter";
-import ToDataProcessor from "./ToDataProcessor";
+import ToDataProcessor from "../ToDataProcessor";
 import ObservableMixin, { Observable } from "@ckeditor/ckeditor5-utils/src/observablemixin";
 import mix from "@ckeditor/ckeditor5-utils/src/mix";
