+++ packages/ckeditor5-coremedia-richtext/src/rules/Schema.ts
@@ -5,7 +5,7 @@ import { TextFilterParams } from "@coremedia/ckeditor5-dataprocessor-support/Tex
 import RichTextSchema, { Strictness } from "../RichTextSchema";
 import { ElementsFilterRuleSetConfiguration } from "@coremedia/ckeditor5-dataprocessor-support/Rules";
 import { DataProcessor } from "@ckeditor/ckeditor5-engine/src/dataprocessor/dataprocessor";
-import RichTextDataProcessor from "../RichTextDataProcessor";
+import RichTextDataProcessor from "../processors/RichTextDataProcessor";
 
 export const defaultSchema = new RichTextSchema(Strictness.STRICT);
